+++ delta/app/src/main/scala/ch/epfl/bluebrain/nexus/delta/routes/ResourcesRoutes.scala
@@ -189,7 +189,7 @@ final class ResourcesRoutes(
                                   emit(
                                     Created,
                                     resources
-                                      .tag(id, ref, schemaOpt, tag, tagRev.toInt, rev)
+                                      .tag(id, ref, schemaOpt, tag, tagRev, rev)
                                       .tapEval(index(ref, _, mode))
                                       .map(_.void)
                                       .rejectWhen(wrongJsonOrNotFound)
