+++ delta/plugins/blazegraph/src/main/scala/ch/epfl/bluebrain/nexus/delta/plugins/blazegraph/BlazegraphViews.scala
@@ -238,7 +238,7 @@ final class BlazegraphViews(
       state   <- id match {
                    case Latest(_)        => log.stateOr(project, iri, notFound)
                    case Revision(_, rev) =>
-                     log.stateOr(project, iri, rev.toInt, notFound, RevisionNotFound)
+                     log.stateOr(project, iri, rev, notFound, RevisionNotFound)
                    case Tag(_, tag)      =>
                      log.stateOr(project, iri, tag, notFound, TagNotFound(tag))
                  }
@@ -325,7 +325,7 @@ object BlazegraphViews {
     * Constructs a projectionId for a blazegraph view
     */
   def projectionId(view: IndexingViewResource): String =
-    projectionId(view.value.uuid, view.rev.toInt)
+    projectionId(view.value.uuid, view.rev)
 
   /**
     * Constructs a projectionId for a blazegraph view
@@ -337,7 +337,7 @@ object BlazegraphViews {
     * Constructs the namespace for a Blazegraph view
     */
   def namespace(view: IndexingViewResource, prefix: String): String =
-    namespace(view.value.uuid, view.rev.toInt, prefix)
+    namespace(view.value.uuid, view.rev, prefix)
 
   /**
     * Constructs the namespace for a Blazegraph view
